refactor: remove TEMPLATE
